better encapsulation for `AttachInstructionsAction`
This PR brings the menu- and keybinding-regirstrion into the action itself and avoids the "external" use of `registerCommandAndKeybindingRule` and `appendMenuItem`. With that it avoids the need for `runAttachInstructionsAction` which was an indirection to invoke the action itself.
